pyproject.toml
[project]
name = "Blaze Emulator"
version = "0.0.1"
authors = [
  { name="Matias Espinoza", email="contact@krozt.dev" },
]
description = "A lightweight Habbo emulator written in Python, dedicated to replicating the Habbo R18 version."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python_dotenv",
]

[project.urls]
"Homepage" = "https://github.com/KrozT/blaze-emulator"
"Bug Tracker" = "https://github.com/KrozT/blaze-emulator/issues"